requirements.txt
cycler==0.10.0
easyocr==1.4.1
fire==0.4.0
imageio==2.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python-headless==4.5.3.56
Pillow==8.2.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
termcolor==1.1.0
tifffile==2021.8.30
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
